refactor(workspace): move tool-names to constants directory
Restructured from workspace/tool-names.ts to workspace/constants/index.ts
to follow the wildcard export pattern in package.json (./* -> ./dist/*/index.js).

- Moved tool name constants to workspace/constants/index.ts
- Added browser-safe export comment
- Updated all imports to use @mastra/core/workspace/constants
- Removed explicit package.json export (wildcard handles it)
- Removed entry from tsup.config.ts
